requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
alembic==1.13.1; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
apache-airflow[amazon]==2.9.0; python_version < '3.13' and python_version ~= '3.8'
apache-airflow-providers-amazon==8.20.0
apache-airflow-providers-common-io==1.3.1; python_version ~= '3.8'
apache-airflow-providers-common-sql==1.12.0; python_version ~= '3.8'
apache-airflow-providers-fab==1.0.3; python_version ~= '3.8'
apache-airflow-providers-ftp==3.8.0; python_version ~= '3.8'
apache-airflow-providers-http==4.10.1; python_version ~= '3.8'
apache-airflow-providers-imap==3.5.0; python_version ~= '3.8'
apache-airflow-providers-smtp==1.6.1; python_version ~= '3.8'
apache-airflow-providers-sqlite==3.7.1; python_version ~= '3.8'
apispec[yaml]==6.6.0; python_version >= '3.8'
argcomplete==3.3.0; python_version >= '3.8'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
asn1crypto==1.5.1
attrs==23.2.0; python_version >= '3.7'
babel==2.14.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
blinker==1.7.0; python_version >= '3.8'
boto3==1.34.85; python_version >= '3.8'
botocore==1.34.85; python_version >= '3.8'
cachelib==0.9.0; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
clickclick==20.10.2
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
colorlog==4.8.0
configupdater==3.2; python_version >= '3.6'
connexion[flask]==2.14.2; python_version >= '3.6'
cron-descriptor==1.4.3
croniter==2.0.3; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
cryptography==42.0.5; python_version >= '3.7'
deprecated==1.2.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dill==0.3.8; python_version >= '3.8'
dnspython==2.6.1; python_version >= '3.8'
docutils==0.21.1; python_version >= '3.9'
email-validator==2.1.1; python_version >= '3.8'
flask==2.2.5; python_version >= '3.7'
flask-appbuilder==4.4.1; python_version ~= '3.7'
flask-babel==2.0.0
flask-caching==2.1.0; python_version >= '3.7'
flask-jwt-extended==4.6.0; python_version >= '3.7' and python_version < '4'
flask-limiter==3.5.1; python_version >= '3.8'
flask-login==0.6.3; python_version >= '3.7'
flask-session==0.5.0; python_version >= '3.7'
flask-sqlalchemy==2.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-wtf==1.2.1; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2024.3.1; python_version >= '3.8'
google-re2==1.1; python_version ~= '3.8'
googleapis-common-protos==1.63.0; python_version >= '3.7'
grpcio==1.62.1; python_version >= '3.7'
gunicorn==22.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
importlib-metadata==7.0.0; python_version >= '3.8'
importlib-resources==6.4.0; python_version >= '3.8'
inflection==0.5.1; python_version >= '3.5'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.3; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
jsonpath-ng==1.6.1
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
lazy-object-proxy==1.10.0; python_version >= '3.8'
limits==3.10.1; python_version >= '3.7'
linkify-it-py==2.0.3; python_version >= '3.7'
lockfile==0.12.2
lxml==5.2.1; python_version >= '3.6'
mako==1.3.3; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.21.1; python_version >= '3.8'
marshmallow-oneofschema==3.1.1; python_version >= '3.8'
marshmallow-sqlalchemy==0.28.2; python_version >= '3.7'
mdit-py-plugins==0.4.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
minio==7.2.5
more-itertools==10.2.0; python_version >= '3.8'
multidict==6.0.5; python_version >= '3.7'
opentelemetry-api==1.24.0; python_version >= '3.8'
opentelemetry-exporter-otlp==1.24.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-common==1.24.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-grpc==1.24.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-http==1.24.0; python_version >= '3.8'
opentelemetry-proto==1.24.0; python_version >= '3.8'
opentelemetry-sdk==1.24.0; python_version >= '3.8'
opentelemetry-semantic-conventions==0.45b0; python_version >= '3.8'
ordered-set==4.1.0; python_version >= '3.7'
packaging==24.0; python_version >= '3.7'
pathspec==0.12.1; python_version >= '3.8'
pendulum==3.0.0; python_version >= '3.8'
pluggy==1.4.0; python_version >= '3.8'
ply==3.11
prison==0.2.1
protobuf==4.25.3; python_version >= '3.8'
psutil==5.9.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyathena==3.6.0; python_full_version >= '3.8.1'
pycparser==2.22; python_version >= '3.8'
pycryptodome==3.20.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pygments==2.17.2; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-daemon==3.0.1; python_version >= '3'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-nvd3==0.15.0
python-slugify==8.0.4; python_version >= '3.7'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
redshift-connector==2.1.1; python_version >= '3.6'
referencing==0.34.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rich==13.7.1; python_full_version >= '3.7.0'
rich-argparse==1.4.0; python_version >= '3.7'
rpds-py==0.18.0; python_version >= '3.8'
s3transfer==0.10.1; python_version >= '3.8'
scramp==1.4.5; python_version >= '3.8'
setproctitle==1.3.3; python_version >= '3.7'
setuptools==69.5.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
sqlalchemy==1.4.52; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy-jsonfield==1.0.2; python_full_version >= '3.7.0'
sqlalchemy-redshift==0.8.14; python_version >= '3.4'
sqlalchemy-utils==0.41.2; python_version >= '3.7'
sqlparse==0.5.0; python_version >= '3.8'
tabulate==0.9.0; python_version >= '3.7'
tenacity==8.2.3; python_version >= '3.7'
termcolor==2.4.0; python_version >= '3.8'
text-unidecode==1.3
time-machine==2.14.1; implementation_name != 'pypy'
typing-extensions==4.11.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
uc-micro-py==1.0.3; python_version >= '3.7'
unicodecsv==0.14.1
universal-pathlib==0.2.2; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
watchtower==3.1.0; python_version >= '3.7'
werkzeug==2.2.3; python_version >= '3.7'
wrapt==1.16.0; python_version >= '3.6'
wtforms==3.1.2; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
zipp==3.18.1; python_version >= '3.8'
